Obtain landcover from GetFeatureInfo
Concludes the effort to obtain the landcover  information upon clicking
the map. This included fixing the map to show the landcover overlay for
both Ordnance Survey and Open Street Map layers. In the process, Leaflet
was upgraded and the OS resolutions tinkered with, allowing zoom levels
for both map projections to match.